max flow problem
It is an implementation of max flow problem in python. It is a way of minimizing the cost required to deliver maximum amount of flow possible in the network. It can be said as an extension of maximum flow problem with an added constraint on cost (per unit flow) of flow for each edge.